ai-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6
google-generativeai==0.3.2
pillow==10.2.0
pytesseract==0.3.10
presidio-analyzer==2.2.33
presidio-anonymizer==2.2.33
google-cloud-vision==3.5.0
boto3==1.34.34
httpx==0.26.0
numpy==1.26.3
opencv-python==4.9.0.80


